Bug 1988947 - Fix setAppLanguageDifferentThanSystemLanguageTest UI test r=ajoltan
The UI test was flaky because it was accessing the language list view before it was available and on a different occasion the test failed to access the “Languages” sub menu and opened the “Accessibility” sub menu instead.  Added a couple of 1s waits for the app window to be updated for better syncing before opening the “Languages” sub-menu and then before accessing the languages list.
The UI test successfully passed 50x on Firebase

Try: [[ https://treeherder.mozilla.org/jobs?repo=try&landoCommitID=169518 | link ]]

Differential Revision: https://phabricator.services.mozilla.com/D276832