Bug 2003682 - Optimize the thumbnail size calculation of the tab grid thumbnails r=android-reviewers,jdelorenzo
During the redesign, we needed some more robust thumbnail calculations in order to support the updated design specs. This led to the individual tab grid items being wrapped with `BoxWithConstraints`, so each item could get its precise measurement to calculate the thumbnail size. However, this led to a performance degradation when combining this into a `LazyVerticalGrid`, as these measurements were occurring frequently and redundantly, as all grid items were of equal size.

To solve for this, the parent `LazyVerticalGrid` is now wrapped by the `BoxWithConstraints`, and the measurement happens once for _all_ tab items.

Differential Revision: https://phabricator.services.mozilla.com/D276571